Add separate component for results list
