+++ react/app-main/components/SearchResultsList/index.ts
@@ -1,0 +1,1 @@
+export { SearchResultsList } from './SearchResultsList';
