+++ react/app-main/components/SearchResultsPage/SearchResultsPage.tsx
@@ -1,40 +1,19 @@
 import * as React from 'react';
 import { useParams } from 'react-router';
 import { LoadingSpinner } from 'tmw-main/components/LoadingSpinner';
-import { ResourceTile } from 'tmw-main/components/ResourceTile';
 import { Resource } from 'tmw-main/constants/app-types';
 import { ajaxPost } from 'tmw-common/utils/ajax';
 import { serializeResourcesFromAPI } from 'tmw-main/utils/api-serialize';
 import { parseSearchTags } from 'tmw-main/utils/tags-search-url';
 import { useViewport } from 'tmw-common/components/ViewportProvider';
 import { VIEWPORT_BREAKPOINTS } from 'tmw-main/constants/app-constants';
+import { SearchResultsList } from 'tmw-main/components/SearchResultsList';
 
 import './search-results-page.less';
 
 export const SearchResultsPage: React.FunctionComponent = () => {
     const { width } = useViewport();
     const isMobileViewport = width < VIEWPORT_BREAKPOINTS.MOBILE;
-
-    const resultsListElement = React.useRef<HTMLDivElement>(null);
-    const [scrollLeftPosition, setScrollLeftPosition] = React.useState(0);
-    const [scrollRightPosition, setScrollRightPosition] = React.useState(width * 0.88);
-
-    const onResultsListScroll = (e: React.UIEvent<HTMLElement>): void => {
-        const { scrollLeft, scrollWidth, clientWidth } = e.currentTarget;
-        setScrollLeftPosition(scrollLeft);
-        setScrollRightPosition(scrollWidth - clientWidth - scrollLeft);
-    };
-
-    const onRightArrowClick = (): void => {
-        resultsListElement?.current?.scrollBy({ left: 360, behavior: 'smooth' });
-    };
-
-    const onLeftArrowClick = (): void => {
-        resultsListElement?.current?.scrollBy({ left: -360, behavior: 'smooth' });
-    };
-
-    const showLeftArrow = scrollLeftPosition > 0;
-    const showRightArrow = scrollRightPosition > 0;
 
     const [isLoading, setIsLoading] = React.useState<boolean>(true);
     const [resultResources, setResultResources] = React.useState<Resource[]>([]);
@@ -79,40 +58,7 @@ export const SearchResultsPage: React.FunctionComponent = () => {
                         <div className="search-results-page__top-spacing" />
                     ) : null}
                     <p className="search-results-page__title">{pageTitle}</p>
-                    {!isMobileViewport ? (
-                        <>
-                            {showLeftArrow ? (
-                                <div className="search-results-page__scroll-arrow search-results-page__scroll-arrow-left ">
-                                    <img
-                                        src="/images/chevron-down.svg"
-                                        alt="Scroll left"
-                                        className="search-results-page__scroll-arrow-left-icon"
-                                        onClick={onLeftArrowClick}
-                                    />
-                                </div>
-                            ) : null}
-
-                            {showRightArrow ? (
-                                <div className="search-results-page__scroll-arrow search-results-page__scroll-arrow-right">
-                                    <img
-                                        src="/images/chevron-down.svg"
-                                        alt="Scroll right"
-                                        className="search-results-page__scroll-arrow-right-icon"
-                                        onClick={onRightArrowClick}
-                                    />
-                                </div>
-                            ) : null}
-                        </>
-                    ) : null}
-                    <div
-                        ref={resultsListElement}
-                        onScroll={onResultsListScroll}
-                        className="search-results-page__results-list"
-                    >
-                        {resultResources.map(resource => (
-                            <ResourceTile key={resource.id} resource={resource} />
-                        ))}
-                    </div>
+                    <SearchResultsList resultsList={resultResources} />
                 </>
             )}
         </div>
