+++ react/app-main/components/SearchResultsPage/search-results-page.less
@@ -28,69 +28,4 @@
             text-align: center;
         }
     }
-
-    &__scroll-arrow {
-        position: absolute;
-        z-index: 10;
-        margin-top: 160px;
-        width: 40px;
-        height: 40px;
-    }
-
-    &__scroll-arrow-left {
-        left: 2%;
-    }
-
-    &__scroll-arrow-right {
-        right: 2%;
-    }
-
-    &__scroll-arrow-left-icon {
-        transform: rotate(90deg);
-        filter: drop-shadow( 0px 0px 3px rgba(0, 0, 0, .7));
-        cursor: pointer;
-    }
-
-    &__scroll-arrow-right-icon {
-        transform: rotate(-90deg);
-        filter: drop-shadow( 0px 0px 3px rgba(0, 0, 0, .7));
-        cursor: pointer;
-    }
-
-    &__results-list {
-        margin-top: 20px;
-
-        overflow-x: scroll;
-        overflow-y: hidden;
-        white-space: nowrap;
-
-        margin-left: calc(-6% / 0.88);
-        margin-right: calc(-6% / 0.88);
-        padding-left: calc(6% / 0.88 - 20px);
-        padding-right: calc(6% / 0.88 - 20px);
-        mask-image: linear-gradient(90deg, rgba(0,0,0,0.2) 0%, rgba(0,0,0,0.9) 5%, rgba(0,0,0,1) 6%, rgba(0,0,0,1) 94%, rgba(0,0,0,0.9) 95%, rgba(0,0,0,0.2) 100%);
-        -webkit-mask-image: -webkit-gradient(linear, left top, right top, color-stop(0%, rgba(0,0,0,0.2)), color-stop(5%, rgba(0,0,0,0.9)), color-stop(6%, rgba(0,0,0,1)), color-stop(94%, rgba(0,0,0,1)), color-stop(95%, rgba(0,0,0,0.9)), color-stop(100%, rgba(0,0,0,0.2)));
-
-        scrollbar-width: none;
-
-        &::-webkit-scrollbar {
-            display: none;
-        }
-
-        @media (max-width: @app-viewport-breakpoint-mobile) {
-            overflow-x: hidden;
-            overflow-y: scroll;
-
-            margin-left: 0;
-            margin-right: 0;
-            padding-left: 0;
-            padding-right: 0;
-            mask-image: none;
-            -webkit-mask-image: none;
-
-            display: flex;
-            flex-direction: column;
-            align-items: center;
-        }
-    }
 }
